// src/types/api.ts
export interface PageParams {
    pageNum: number
    pageSize?: number
}

export namespace Login {
    export interface params {
        userName: string
        userPwd: string
    }
}
// 用户管理
export namespace User {
    export interface Params extends PageParams {
        userId?: number
        userName?: string
        state?: number
    }

    export interface SearchParams {
        userId?: number
        userName?: string
        state?: number
    }
    export interface UserItem {
        _id: string
        userId: number
        userName: string
        userEmail: string
        deptId: string
        state: number
        mobile: string
        job: string
        role: number
        roleList: string
        createId: number
        deptName: string
        userImg: string
    }
    export interface CreateParams {
        userName: string
        userEmail: string
        mobile?: number
        deptId: string
        job?: string
        state?: number
        roleList: string[]
        userImg: string
    }
    export interface EditParams extends CreateParams {
        userId: number
    }
}